requirements.txt
PySide6
PyOpenGL
numpy
PyGLM
Pillow
imgui
